add 添加搜索门面 add 添加多数据源，使用设计模式进行优化

+++ src/main/java/com/yanweiyi/misearchbackend/service/impl/PostServiceImpl.java
@@ -297,8 +297,8 @@ public class PostServiceImpl extends ServiceImpl<PostMapper, Post> implements Po
 
     @Override
     public Page<PostVO> listPostVOByPage(PostQueryRequest postQueryRequest, HttpServletRequest request) {
-        int current = postQueryRequest.getCurrent();
-        int size = postQueryRequest.getPageSize();
+        long current = postQueryRequest.getCurrent();
+        long size = postQueryRequest.getPageSize();
         Page<Post> postPage = this.page(new Page<>(current, size), this.getQueryWrapper(postQueryRequest));
         return this.getPostVOPage(postPage, request);
     }
